2025_Advanced_Earth_Observation/requirements.txt
affine==2.4.0
attrs==25.3.0
certifi==2025.4.26
click==8.2.0
click-plugins==1.1.1
cligj==0.7.2
numpy==2.2.6
pyparsing==3.2.3
rasterio==1.4.3
affine==2.4.0
attrs==25.3.0
certifi==2025.4.26
click==8.2.0
click-plugins==1.1.1
cligj==0.7.2
numpy==2.2.6
pandas==2.2.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
rasterio==1.4.3
six==1.17.0
tzdata==2025.2
affine==2.4.0
attrs==25.3.0
certifi==2025.4.26
click==8.2.0
click-plugins==1.1.1
cligj==0.7.2
numpy==2.2.6
pandas==2.2.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
rasterio==1.4.3
six==1.17.0
tzdata==2025.2
affine==2.4.0
attrs==25.3.0
certifi==2025.4.26
click==8.2.0
click-plugins==1.1.1
cligj==0.7.2
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
pandas==2.2.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
rasterio==1.4.3
six==1.17.0
tzdata==2025.2
affine==2.4.0
attrs==25.3.0
certifi==2025.4.26
click==8.2.0
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
rasterio==1.4.3
six==1.17.0
tzdata==2025.2
affine==2.4.0
attrs==25.3.0
certifi==2025.4.26
click==8.2.0
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
rasterio==1.4.3
six==1.17.0
tzdata==2025.2
